Fixed ALLOCATE for 64-bit argument (previously truncated to 32-bits).
 Changes to be committed:
	modified:   ForthVM.cpp

+++ src/ForthVM.cpp
@@ -1759,7 +1759,7 @@ int CPP_allocate()
     return E_V_NOT_IVAL;  // need an int
 #endif
 
-  unsigned int requested = TOS;
+  unsigned long requested = TOS;
   byte *p = new (nothrow) byte[requested];
   PUSH_ADDR( (long int) p )
   PUSH_IVAL( p ? 0 : -1 )
